VLA4Desk Franka：仓库内默认路径、LoRA 实验与 LIBERO .pt 冷启动
统一 output_paths（datasets/hf_cache/checkpoints/vla4desk），文档与脚本默认相对仓库根；gitignore 仅跟踪 vla4desk/*.txt。新增 Franka 135 demos / LoRA 实验、consolidated .pt 预加载、T5 fp32 与 encoder-only 导出。

+++ cosmos_policy/datasets/t5_embedding_utils.py
@@ -33,7 +33,7 @@ def generate_t5_embeddings(
     max_gpu_mem_gib: Optional[float] = None,
     cache_dir: Optional[str] = None,
     local_files_only: Optional[bool] = None,
-    dtype: str = "bf16",
+    dtype: str = "fp32",
 ) -> Dict[str, torch.Tensor]:
     """
     Generate T5 text embeddings for a list of commands.
